requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.8
packaging==24.1
pillow==10.4.0
Python_IO==0.3
requests==2.32.3
urllib3==2.2.2
